requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beanie==2.0.1
bidict==0.23.1
click==8.3.1
cross-web==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
graphql-core==3.2.8
h11==0.16.0
idna==3.11
lazy-model==0.4.0
motor==3.7.1
packaging==26.0
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
six==1.17.0
starlette==0.52.1
strawberry-graphql==0.311.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wsproto==1.3.2
